[Feat] : 키변환 클릭시 로딩 UI 구현
