+++ ChordLululala/Views/ScoreViews/ScoreView.swift
@@ -18,19 +18,21 @@ struct ScoreView : View {
             /// 악보 헤더부분
             if !viewModel.playmodeViewModel.isOn {
                 ScoreHeaderView(viewModel: viewModel.headerViewModel,
-                                annotationVM: viewModel.annotationViewModel
+                                annotationVM: viewModel.annotationViewModel,
+                                isTransposing: viewModel.isTransposingViewModel
                 )
-                    .transition(.move(edge: .top).combined(with: .opacity)) // 슬라이드 인아웃 + 페이드 효과
+                .transition(.move(edge: .top).combined(with: .opacity)) // 슬라이드 인아웃 + 페이드 효과
             }
-                
+            
             /// 악보 바디 뷰
             ScoreMainBodyView(
                 pdfViewModel: viewModel.pdfViewModel,
                 playmodeViewModel: viewModel.playmodeViewModel,
                 pageNavViewModel: viewModel.pageNavViewModel,
-                annotationVM: viewModel.annotationViewModel
+                annotationVM: viewModel.annotationViewModel,
+                isTransposing: viewModel.isTransposingViewModel
             )
-        
+            
         }
         .navigationBarHidden(true)
